Started Implementing XY Plot for Zreal/imag data
Since still waiting for access to circuits lab for current draw test and still waiting on Android for BLE testing, decided to deal with data visualization. Managed to get project to use the androidplot library and was able to add an empty plot to the main layout. Need to figure out Kotlin code for the activity to actually add points to the plot.