redrawNyquist
Wrote a function called redrawNyquist. When called, it clears the graph and draws a new one based off current Zreal/Zimag values. Once BLE is working, this will be called twice:
(1) when the start button is clicked. This just clears Zreal and Zimag, making the graph a clear plot.
(2) when all of the BLE data is received. This will then draw the Nyquist plot for the most recent sample.

NOTE: I am electing to clear at the beginning of a sample so that the user can more easily tell when a sample is running (clear) vs when a sample is finished (semicircle).