Semi-AutoConnect for App
On startup, if user clicks on "Connect", the device will automatically start & stop BLE scan before connecting to HELPStat. In previous versions, this was a 3-step process.

Tried to have the app immediately connect on startup, but causes the app to crash if BLE is not enabled (instead of prompting the user to turn on bluetooth).

Also did some color changes to make it look better. I don't think it really worked.